Add stage2 module
This stage serves the role of a lexer. After this pass
all the whitespace should be insignificant.
This stage also differentiates between various types
of identifiers – commandlets, variables, flags and field names.

Add a pretty-printed debug output of the parser
enabled by a flag for analyze subcommand.

The parser v2 is now being run, although its results are ignored.