Add ItemKind enum to differentiate classes and functions
So now, every usage of name (which assumed a function)
is replaced with a new Item struct, which contains both
a name and ItemKind.

The Item struct can be both owned and non-owned,
with case sensitive and insensitive variants.

The alternative would be to split items in Scope
to handle functions, classes etc. separately.
Decided to not do it.
Another alternative would be to create Item enum,
with names as variant payloads.